+++ src/apis/tale/ManageTaleAPICalls.js
